+++ app/tests/test_recommend.py
@@ -95,16 +95,10 @@ class RecommendTest(TestCase):
 
     # ------------------------------------------------------------------------------------------------------------------
     def tearDown(self):
-        books = Book.objects.all()
-
-        for book in books:
-            self.clean_book_file(book)
+        for book in Book.objects.all():
+            if os.path.exists(book.book_file.path):
+                os.remove(book.book_file.path)
+            if book.photo and os.path.exists(book.photo.path):
+                os.remove(book.photo.path)
 
         recommend.START_RECOMMEND = 10
-
-    # ------------------------------------------------------------------------------------------------------------------
-    def clean_book_file(self, book):
-        os.remove(book.book_file.path)
-
-        if book.photo and os.path.exists(book.photo.path):
-            os.remove(book.photo.path)
